Django演習-様々なビューの作成 Django演習3のプルリクエストを上げさせて頂きます。
本章では
・"index" ページ
・"detail" ページ
・"result" ページ
・"vote"ページ
・例外処理
をDjangoで作成しました。

今回確認していただきたい内容は二点あります。

(i) indexページが表示されること

(ii) detailページに飛んだ際、404エラーが発生すること(与える数を引数にとるオブジェクトが存在しないため)

確認していただく手順は以下の通りです。

1）python manage.py runserver をターミナルで入力し、サーバーを起動

2）＜(i)の確認＞ "http://127.0.0.1:8000/polls/"を開き、箇条書きの質問のリストが表示されていることを確認（"・What's up?"という項目）

3）＜(ii)の確認＞
"http://127.0.0.1:8000/polls/34"を開き、404のエラーが表示されている事を確認（No Question matches the given query.というもの）

以上の手順でご確認をお願いいたします。
