+++ django/mysite/polls/admin.py
@@ -1,6 +1,6 @@
+# Register your models here.
 from django.contrib import admin
 
-# Register your models here.
 from .models import Question
 
 admin.site.register(Question)
