+++ django/mysite/polls/models.py
@@ -1,15 +1,17 @@
+# Create your models here.
 import datetime
 
 from django.db import models
 from django.utils import timezone
 
-# Create your models here.
 
 class Question(models.Model):
     question_text = models.CharField(max_length=200)
     pub_date = models.DateTimeField("date published")
     def __str__(self):
         return self.question_text
+    def was_published_recently(self):
+        return self.pub_date >= timezone.now() - datetime.timedelta(days=1)
 
 
 class Choice(models.Model):
